[IMP] mail_broker_whatsapp: Pass logic to components
