+++ mail_broker_whatsapp/__init__.py
@@ -1,2 +1,3 @@
 from . import models
+from . import services
 from . import wizards
