+++ mail_broker_whatsapp/models/__init__.py
@@ -1,4 +1,2 @@
-from . import mail_broker_channel
-from . import mail_message_broker
 from . import mail_broker
 from . import mail_thread
