+++ mail_broker_whatsapp/services/__init__.py
@@ -1,0 +1,1 @@
+from . import mail_broker_service
